Removed the unnecessary base class from 400
